introduce internal getVariableSignature utility
extracted from my fragment arguments scratch branch
